pyproject.toml
[project]
name = "assigment-ticket-extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.2",
    "openai>=2.8.1",
    "pydantic>=2.12.4",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
